acc: Refactor starting test server in a separate file
